restructure nextjs project and fix tsconfig
