+++ client/src/pages/_app.tsx
@@ -1,0 +1,10 @@
+import Layout from "@/components/layout";
+import type { AppProps } from "next/app";
+
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  );
+}
